requirements.txt
Django >= 3.2.7,<=3.2.99
dj-rest-auth >= 2.1.11
django-allauth >= 0.45.0
django-cors-headers >= 3.10.0
djangorestframework >= 3.12.4
django-seed >= 0.3.1
drf-yasg >= 1.20.0
drf-url-filters >= 0.5.1
python-dotenv >= 0.19.1
psycopg2 >= 2.9.1
